backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
beautifulsoup4
lxml
